python-service/requirements.txt
fastapi
uvicorn[standard]
llama-index-core
llama-index-embeddings-huggingface
llama-index-vector-stores-mongodb
pymongo
sentence-transformers
torch
transformers
python-dotenv
pypdf
pdfplumber
pymupdf